logic refactor for matching main
