+++ backend/src/main/java/chess/vmsb/Application.java
@@ -1,0 +1,13 @@
+package chess.vmsb;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Application {
+
+public static void main(String[] args) {
+SpringApplication.run(Application.class, args);
+}
+
+}
