+++ backend/src/test/java/chess/vmsb/ApplicationTests.java
@@ -1,4 +1,4 @@
-package vmsb.chess.logic;
+package chess.vmsb;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -6,8 +6,8 @@ import org.springframework.boot.test.context.SpringBootTest;
 @SpringBootTest
 class ApplicationTests {
 
-	@Test
-	void contextLoads() {
-	}
+    @Test
+    void contextLoads() {
+    }
 
 }
